refactor(data): move load function into designated modules
feat(figure)!: change fig1 and soon make fig2 redundant
