+++ src/paper1_code/load/__init__.py
@@ -1,0 +1,23 @@
+"""Initialize the load module."""
+
+from paper1_code.load import (
+    cesm2,
+    g16,
+    j05,
+    m20,
+    ob16,
+    pinatubo,
+    t10,
+    tambora,
+)
+
+__all__ = [
+    "cesm2",
+    "g16",
+    "j05",
+    "m20",
+    "ob16",
+    "pinatubo",
+    "t10",
+    "tambora",
+]
