+++ src/paper1_code/scripts/__init__.py
@@ -1,5 +1,1 @@
 """Initialize the scripts module."""
-
-from paper1_code.scripts import if_save
-
-__all__ = ["if_save"]
