+++ src/paper1_code/scripts/gen_fig1.py
@@ -7,7 +7,6 @@ import cosmoplots
 import matplotlib as mpl
 import numpy as np
 import scipy
-import xarray as xr
 from matplotlib import patches as mpatches
 from matplotlib import pyplot as plt
 from matplotlib.legend_handler import HandlerLine2D
@@ -23,77 +22,14 @@ class SetupNeededData:
     """Class that loads all data used in the plotting procedures."""
 
     def __init__(self):
-        files = self._load_all_files()
-        self.medium, self.plus, self.strong = self._get_shifted_data(files)
-
-    def _load_all_files(self) -> core.utils.load_auto.FindFiles:
-        """Make sure that all necessary files are available."""
-        matching_files = (
-            core.utils.load_auto.FindFiles(ft=".npz")
-            .find(
-                "e_BWma1850",
-                [f"ens{i+1}" for i in range(4)],
-                {"medium", "medium-plus", "strong"},
-                "TREFHT",
-                "h1",
-            )
-            .sort("sim", "attr", "ensemble")
-        )
-        m_list = [
-            ("e_BWma1850", "medium", "ens1", "TREFHT", "h1", "20230828"),
-            ("e_BWma1850", "medium", "ens2", "TREFHT", "h1", "20230828"),
-            ("e_BWma1850", "medium", "ens3", "TREFHT", "h1", "20230828"),
-            ("e_BWma1850", "medium", "ens4", "TREFHT", "h1", "20230828"),
-            ("e_BWma1850", "medium-plus", "ens1", "TREFHT", "h1", "20230828"),
-            ("e_BWma1850", "medium-plus", "ens2", "TREFHT", "h1", "20230828"),
-            ("e_BWma1850", "medium-plus", "ens3", "TREFHT", "h1", "20230828"),
-            ("e_BWma1850", "medium-plus", "ens4", "TREFHT", "h1", "20230828"),
-            ("e_BWma1850", "strong", "ens1", "TREFHT", "h1", "20230828"),
-            ("e_BWma1850", "strong", "ens2", "TREFHT", "h1", "20230828"),
-            ("e_BWma1850", "strong", "ens3", "TREFHT", "h1", "20230828"),
-            ("e_BWma1850", "strong", "ens4", "TREFHT", "h1", "20230828"),
-        ]
-        if matching_files.get_files() != m_list:
-            # Here we might want to ask the user if they would like to download the files
-            # from the Fram archive storage.
-            raise FileNotFoundError("Missing the nescessary files to do this analysis.")
-        return matching_files
-
-    def _get_shifted_data(
-        self, files: core.utils.load_auto.FindFiles
-    ) -> tuple[list[xr.DataArray], list[xr.DataArray], list[xr.DataArray]]:
-        """Get shifted data from the medium, medium-plus and strong eruption simulations.
-
-        Parameters
-        ----------
-        files : core.utils.load_auto.FindFiles
-            An instance of FindFiles that contains the necessary files
-
-        Returns
-        -------
-        medium : list[xr.DataArray]
-            Shifted data from the medium eruption simulation.
-        medium_plus : list[xr.DataArray]
-            Shifted data from the medium-plus eruption simulation.
-        strong : list[xr.DataArray]
-            Shifted data from the strong eruption simulation.
-        """
-        medium = files.copy().keep("medium").load()
-        plus = files.copy().keep("medium-plus").load()
-        strong = files.copy().keep("strong").load()
-        for i, a in enumerate(medium):
-            medium[i] = a.compute()
-        for i, a in enumerate(plus):
-            plus[i] = a.compute()
-        for i, a in enumerate(strong):
-            strong[i] = a.compute()
-        medium_ = core.utils.time_series.remove_seasonality(medium)
-        plus_ = core.utils.time_series.remove_seasonality(plus)
-        strong_ = core.utils.time_series.remove_seasonality(strong)
-        medium__ = core.utils.time_series.shift_arrays(medium_)
-        plus__ = core.utils.time_series.shift_arrays(plus_)
-        strong__ = core.utils.time_series.shift_arrays(strong_)
-        return medium__, plus__, strong__
+        (
+            self.medium,
+            self.plus,
+            self.strong,
+            *_,
+            # ) = core.load.cesm2.get_aod_arrs(remove_seasonality=True, shift=0)
+        ) = core.load.cesm2.get_rf_arrs(remove_seasonality=True, shift=0)
+        # ) = core.load.cesm2.get_trefht_arrs(remove_seasonality=True)
 
 
 class DoPlotting:
@@ -102,6 +38,7 @@ class DoPlotting:
     def __init__(self, print_summary: bool):
         self.print_summary = print_summary
         self.data = SetupNeededData()
+        self.n_year = 6
 
     def waveform_max(self) -> mpl.figure.Figure:
         """Compare shape of TREFHT variable from medium and strong eruption simulations.
@@ -120,14 +57,12 @@ class DoPlotting:
             strong[i] = s[: len(medium[0])]
 
         # Find median values
-        eq_temp = core.config.MEANS["TREFHT"]
-        medium_med = np.median(medium, axis=0) - eq_temp
-        plus_med = np.median(plus, axis=0) - eq_temp
-        strong_med = np.median(strong, axis=0) - eq_temp
-        win_len = 365
-        medium_max = -scipy.signal.savgol_filter(medium_med, win_len, 3).min()
-        plus_max = -scipy.signal.savgol_filter(plus_med, win_len, 3).min()
-        strong_max = -scipy.signal.savgol_filter(strong_med, win_len, 3).min()
+        medium_med = np.median(medium, axis=0)
+        plus_med = np.median(plus, axis=0)
+        strong_med = np.median(strong, axis=0)
+        medium_max = scipy.signal.savgol_filter(medium_med, self.n_year, 3).max()
+        plus_max = scipy.signal.savgol_filter(plus_med, self.n_year, 3).max()
+        strong_max = scipy.signal.savgol_filter(strong_med, self.n_year, 3).max()
         medium_scaled = medium_med / medium_max
         plus_scaled = plus_med / plus_max
         strong_scaled = strong_med / strong_max
@@ -136,17 +71,17 @@ class DoPlotting:
         n = 1
         low_range = np.linspace(MIN_PERCENTILE, 50, num=n, endpoint=False)
         high_range = np.linspace(50, MAX_PERCENTILE, num=n + 1)[1:]
-        medium_perc1 = np.percentile(medium, low_range, axis=0) - eq_temp
+        medium_perc1 = np.percentile(medium, low_range, axis=0)
         medium_perc1 = medium_perc1 / medium_max
-        medium_perc2 = np.percentile(medium, high_range, axis=0) - eq_temp
+        medium_perc2 = np.percentile(medium, high_range, axis=0)
         medium_perc2 = medium_perc2 / medium_max
-        plus_perc1 = np.percentile(plus, low_range, axis=0) - eq_temp
+        plus_perc1 = np.percentile(plus, low_range, axis=0)
         plus_perc1 = plus_perc1 / plus_max
-        plus_perc2 = np.percentile(plus, high_range, axis=0) - eq_temp
+        plus_perc2 = np.percentile(plus, high_range, axis=0)
         plus_perc2 = plus_perc2 / plus_max
-        strong_perc1 = np.percentile(strong, low_range, axis=0) - eq_temp
+        strong_perc1 = np.percentile(strong, low_range, axis=0)
         strong_perc1 = strong_perc1 / strong_max
-        strong_perc2 = np.percentile(strong, high_range, axis=0) - eq_temp
+        strong_perc2 = np.percentile(strong, high_range, axis=0)
         strong_perc2 = strong_perc2 / strong_max
 
         x_m = medium[0].time.data
@@ -169,8 +104,6 @@ class DoPlotting:
         strong_fill = mpatches.Patch(facecolor=CLR[2], alpha=1.0, linewidth=0)
         for p1, p2 in zip(strong_perc1, strong_perc2, strict=True):
             ax.fill_between(x_s, p1, p2, alpha=alpha, color=CLR[2], edgecolor=None)
-        # Vertical line at the time of the eruption
-        ax.axvline(1850 + (31 + 15) / 356, c="k")
 
         # Combine shading and line labels
         plt.legend(
@@ -184,13 +117,14 @@ class DoPlotting:
                 r"C2W$-$, $C" + f" = {plus_max:.2f}" + r"$",
                 r"C2W$\downarrow$, $C" + f" = {medium_max:.2f}" + r"$",
             ],
-            loc="lower right",
+            loc="upper right",
             handler_map={
                 medium_line: HandlerLine2D(marker_pad=0),
                 plus_line: HandlerLine2D(marker_pad=0),
                 strong_line: HandlerLine2D(marker_pad=0),
             },
             framealpha=0.6,
+            fontsize=core.config.FONTSIZE,
         )
         return plt.gcf()
 
@@ -211,13 +145,12 @@ class DoPlotting:
             strong[i] = s[: len(medium[0])]
 
         # Find median values
-        eq_temp = core.config.MEANS["TREFHT"]
-        medium_med = np.median(medium, axis=0) - eq_temp
-        plus_med = np.median(plus, axis=0) - eq_temp
-        strong_med = np.median(strong, axis=0) - eq_temp
-        medium_int = -np.trapz(medium_med, medium[0].time.data)
-        plus_int = -np.trapz(plus_med, plus[0].time.data)
-        strong_int = -np.trapz(strong_med, strong[0].time.data)
+        medium_med = np.median(medium, axis=0)
+        plus_med = np.median(plus, axis=0)
+        strong_med = np.median(strong, axis=0)
+        medium_int = np.trapz(medium_med, medium[0].time.data)
+        plus_int = np.trapz(plus_med, plus[0].time.data)
+        strong_int = np.trapz(strong_med, strong[0].time.data)
         medium_scaled = medium_med / medium_int
         plus_scaled = plus_med / plus_int
         strong_scaled = strong_med / strong_int
@@ -226,17 +159,17 @@ class DoPlotting:
         n = 1
         low_range = np.linspace(MIN_PERCENTILE, 50, num=n, endpoint=False)
         high_range = np.linspace(50, MAX_PERCENTILE, num=n + 1)[1:]
-        medium_perc1 = np.percentile(medium, low_range, axis=0) - eq_temp
+        medium_perc1 = np.percentile(medium, low_range, axis=0)
         medium_perc1 = medium_perc1 / medium_int
-        medium_perc2 = np.percentile(medium, high_range, axis=0) - eq_temp
+        medium_perc2 = np.percentile(medium, high_range, axis=0)
         medium_perc2 = medium_perc2 / medium_int
-        plus_perc1 = np.percentile(plus, low_range, axis=0) - eq_temp
+        plus_perc1 = np.percentile(plus, low_range, axis=0)
         plus_perc1 = plus_perc1 / plus_int
-        plus_perc2 = np.percentile(plus, high_range, axis=0) - eq_temp
+        plus_perc2 = np.percentile(plus, high_range, axis=0)
         plus_perc2 = plus_perc2 / plus_int
-        strong_perc1 = np.percentile(strong, low_range, axis=0) - eq_temp
+        strong_perc1 = np.percentile(strong, low_range, axis=0)
         strong_perc1 = strong_perc1 / strong_int
-        strong_perc2 = np.percentile(strong, high_range, axis=0) - eq_temp
+        strong_perc2 = np.percentile(strong, high_range, axis=0)
         strong_perc2 = strong_perc2 / strong_int
 
         x_m = medium[0].time.data
@@ -259,8 +192,6 @@ class DoPlotting:
         strong_fill = mpatches.Patch(facecolor=CLR[2], alpha=1.0, linewidth=0)
         for p1, p2 in zip(strong_perc1, strong_perc2, strict=True):
             ax.fill_between(x_s, p1, p2, alpha=alpha, color=CLR[2], edgecolor=None)
-        # Vertical line at the time of the eruption
-        ax.axvline(1850 + (31 + 15) / 356, c="k")
 
         # Combine shading and line labels
         plt.legend(
@@ -274,13 +205,14 @@ class DoPlotting:
                 r"C2W$-$, $C" + f" = {plus_int:.2f}" + r"$",
                 r"C2W$\downarrow$, $C" + f" = {medium_int:.2f}" + r"$",
             ],
-            loc="lower right",
+            loc="upper right",
             handler_map={
                 medium_line: HandlerLine2D(marker_pad=0),
                 plus_line: HandlerLine2D(marker_pad=0),
                 strong_line: HandlerLine2D(marker_pad=0),
             },
             framealpha=0.6,
+            fontsize=core.config.FONTSIZE,
         )
         return plt.gcf()
 
@@ -294,7 +226,7 @@ def main(show_output: bool = False):
     wint = plotter.waveform_integrate()
     wmax = plotter.waveform_max()
     if save:
-        SAVE_PATH = core.scripts.if_save.create_savedir()
+        SAVE_PATH = core.utils.if_save.create_savedir()
         wmax.savefig(tmp_dir / "compare-waveform-max")
         wint.savefig(tmp_dir / "compare-waveform-integrate")
         cosmoplots.combine(
