+++ tests/test_config.py
@@ -6,7 +6,7 @@ import paper1_code as core
 
 def test_means() -> None:
     """Test config constants."""
-    finder = core.utils.load_auto.FindFiles()
+    finder = core.utils.find_c2w_files.FindFiles()
     # finder.avail()
     finder.find("e_BWma1850", "control", "TREFHT")
     if len(finder) != 1:
